Add Typescript support
