+++ package.json
@@ -4,7 +4,9 @@
   "description": "",
   "main": "bin/index.js",
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
+    "build": "npm run-script tsc",
+    "test": "echo \"Error: no test specified\" && exit 1",
+    "tsc": "./node_modules/typescript/bin/tsc"
   },
   "author": "Thomas Miller <thomas.d.miller.jr@gmail.com>",
   "license": "ISC",
@@ -13,6 +15,8 @@
   },
   "dependencies": {
     "axios": "^0.19.2",
+    "tslint": "^6.1.2",
+    "typescript": "^3.9.5",
     "yargs": "^15.3.1"
   }
 }
